[feat] Refactor wallet_app/android
