+++ wallet_app/android/app/src/main/StarknetClient.kt
