+++ wallet_app/android/app/src/main/java/com/example/walletapp/ui/activity/AccountPassword.kt
@@ -1,4 +1,4 @@
-package com.example.walletapp
+package com.example.walletapp.ui.activity
 
 import android.app.Activity
 import android.content.Intent
