+++ wallet_app/android/app/src/main/java/com/example/walletapp/ui/activity/ReceiverActivity.kt
@@ -1,4 +1,4 @@
-package com.example.walletapp
+package com.example.walletapp.ui.activity
 
 import android.content.ClipData
 import android.content.ClipboardManager
