+++ wallet_app/android/app/src/main/java/com/example/walletapp/ui/activity/WalletActivity.kt
@@ -1,4 +1,4 @@
-package com.example.walletapp
+package com.example.walletapp.ui.activity
 
 import StarknetClient
 import android.app.Activity
