+++ wallet_app/android/app/src/main/res/theme/Color.kt
